+++ Assets/Scenes/Minigame1.unity
@@ -4320,7 +4320,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2072481279}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 02b849b9427a14751b68372d60391afa, type: 3}
+  m_Script: {fileID: 11500000, guid: bd8df447fd683c344b722d12da038f88, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   ansTime: 15
