--- Assets/Scripts/GameControl.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 02b849b9427a14751b68372d60391afa
