+++ Packages/manifest.json
@@ -4,9 +4,9 @@
     "com.unity.feature.2d": "2.0.1",
     "com.unity.ide.rider": "3.0.31",
     "com.unity.ide.visualstudio": "2.0.22",
-    "com.unity.inputsystem": "1.13.0",
+    "com.unity.inputsystem": "1.13.1",
     "com.unity.multiplayer.center": "1.0.0",
-    "com.unity.render-pipelines.universal": "17.0.3",
+    "com.unity.render-pipelines.universal": "17.0.4",
     "com.unity.test-framework": "1.4.6",
     "com.unity.timeline": "1.8.7",
     "com.unity.ugui": "2.0.0",
