+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 6000.0.39f1
-m_EditorVersionWithRevision: 6000.0.39f1 (15ea7ed0b100)
+m_EditorVersion: 6000.0.43f1
+m_EditorVersionWithRevision: 6000.0.43f1 (97272b72f107)
